Refactored package structure
Split up the classes to have their own representation in the form of a folder/package.
