+++ src/main/java/com/beastlymc/data/heaps/MinHeap.java
@@ -1,14 +1,17 @@
-package com.beastlymc.data.tree.heap;
+package com.beastlymc.data.heaps;
 
 /**
- * MinHeap class represents a min heap data structure where the minimum element is always at the top.
+ * MinHeap class represents a min heap data structure where the minimum element
+ * is always at the top.
  * <p>
- * The class extends the abstract Heap class and provides implementations for its abstract methods.
+ * The class extends the abstract Heap class and provides implementations for
+ * its abstract methods.
  *
- * @param <T> The type of elements stored in the heap. The elements must be comparable with each other.
+ * @param <T> The type of elements stored in the heap. The elements must be
+ *            comparable with each other.
  *
  * @see Comparable
- * @see com.beastlymc.data.tree.heap.Heap
+ * @see Heap
  */
 public class MinHeap<T extends Comparable<T>> extends Heap<T> {
 
@@ -22,7 +25,8 @@ public class MinHeap<T extends Comparable<T>> extends Heap<T> {
     }
 
     /**
-     * Inserts the specified element into the heap and maintains the min-heap property.
+     * Inserts the specified element into the heap and maintains the min-heap
+     * property.
      *
      * @param element The element to be inserted into the heap.
      *
@@ -39,7 +43,8 @@ public class MinHeap<T extends Comparable<T>> extends Heap<T> {
     }
 
     /**
-     * Deletes the minimum element from the heap and maintains the min heap property.
+     * Deletes the minimum element from the heap and maintains the min heap
+     * property.
      *
      * @return The minimum element in the min-heap.
      *
@@ -65,22 +70,26 @@ public class MinHeap<T extends Comparable<T>> extends Heap<T> {
      * @param index The index of the element to be moved.
      */
     private void heapifyUp(int index) {
-        while (hasParent(index) && getParent(index).compareTo(heapArray.get(index)) > 0) {
+        while (hasParent(index) &&
+               getParent(index).compareTo(heapArray.get(index)) > 0) {
             swap(getParentIndex(index), index);
             index = getParentIndex(index);
         }
     }
 
     /**
-     * A helper method to move the element at the root of the heap down until the min-heap property is satisfied
+     * A helper method to move the element at the root of the heap down until
+     * the min-heap property is satisfied
      */
     private void heapifyDown(int index) {
         while (hasLeftChild(index)) {
             int smallerChildIndex = getLeftChildIndex(index);
-            if (hasRightChild(index) && getRightChild(index).compareTo(getLeftChild(index)) < 0) {
+            if (hasRightChild(index) &&
+                getRightChild(index).compareTo(getLeftChild(index)) < 0) {
                 smallerChildIndex = getRightChildIndex(index);
             }
-            if (heapArray.get(index).compareTo(heapArray.get(smallerChildIndex)) < 0) {
+            if (heapArray.get(index)
+                        .compareTo(heapArray.get(smallerChildIndex)) < 0) {
                 break;
             } else {
                 swap(index, smallerChildIndex);
