+++ src/test/java/com/beastlymc/data/heaps/HeapTest.java
@@ -1,4 +1,4 @@
-package com.beastlymc.data.tree.heap;
+package com.beastlymc.data.heaps;
 
 import org.junit.jupiter.api.Test;
