+++ src/test/java/com/beastlymc/data/lists/LinkedListTest.java
@@ -1,4 +1,4 @@
-package com.beastlymc.data;
+package com.beastlymc.data.lists;
 
 import org.junit.jupiter.api.Test;
